MultiCurl without wait for start
